refactor: fixing exception conditionals
