+++ api-geotrack/src/main/java/com/geotrack/apigeotrack/dto/stopoint/FeatureDTO.java
@@ -6,5 +6,5 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record FeatureDTO(@JsonAlias("type") String type,
                          @JsonAlias("properties") PropertiesDTO properties,
-                         @JsonAlias("geometry") GeometryDTO geometry) {
+                         @JsonAlias("zone") GeometryDTO geometry) {
 }
